[Feat][F][COMMON_001] 레이아웃 조정
