+++ frontend/src/assets/main.css
@@ -8,7 +8,8 @@
 /* ── Callog Layout ─────────────────────────────────────── */
 .callog-layout {
   display: flex;
-  min-height: 100vh;
+  height: 100vh;
+  overflow: hidden;
   background: var(--app-bg);
   color: var(--text-primary);
   transition:
@@ -21,10 +22,19 @@
   min-width: 0;
   display: flex;
   flex-direction: column;
+  overflow: hidden;
+}
+
+.callog-body {
+  flex: 1;
+  min-height: 0;
+  display: flex;
+  overflow: hidden;
 }
 
 .callog-content {
   flex: 1;
+  min-width: 0;
   min-height: 0;
   overflow: auto;
   padding: 24px;
